\name{SpatialEffect-package}
\alias{SpatialEffect-package}
\docType{package}
\title{
Spatial Effect
}
\description{
  The function estimate the marginalized individualistic response (MIR) curve defined in Aronow, Samii and Wang (2019) and its standard error.}
\author{
  Cyrus Samii <cds2083@nyu.edu>, New York University   
  Ye Wang <yw1576@nyu.edu>, New York University
}
\details{
  The function estimate the marginalized individualistic response (MIR) curve defined in Aronow, Samii and Wang (2019) and its standard error. 
}

